Rework bin tests as integration tests.
This fixes the issue where `cargo test` did not build the binary before
running the tests.

Cargo assumes that binaries are tested via integration tests, and as
such, only builds the binary and sets `CARGO_BIN_EXE_<name>`, providing
the absolute path to the built binary, while building those integration
tests.

Part of the solution was removing the workspace, by combining the two
packages into a single codebase with the cli code as a feature. I
believe this is the more idiomatic way to implement and lib/bin combo.

Note that the vast majority of the changes are simply moving code
around, and there are only a few changes beyond that.

- `cli/src/` and `core/src/` were merged into a single, top-level `src/`.
- `cli/Cargo.toml` and `core/Cargo.toml` were combined in the top-level
  `Cargo.toml`.
- The cli-specific prod deps were marked as optional.
- A `cli` feature was added that included the cli-specific deps. The
  `cli` feature is enabled by default.
- Most of the binary logic was moved from `src/bin/subg/` to `src/subg/`
  and part of the lib, dependent on the `cli` feature. Most of the
  `pub(crate)` scoping was relaxed to `pub` so it could be used by
  integration tests.
- The internal tests in `src/subg/` where moved to the top-level `tests/`.
- `assert_cmd::Command::cargo_bin` in now passed
  `env!("CARGO_BIN_EXE_subg")`.
- Uses of `right_pad(value,width)` were replaced with
  `format!("{value:<width$}")` as I believe this is a more idiomatic
  approach. This eliminated the need for `src/subg/util.rs`.